add controllers and fix implementation errors

+++ src/main/java/com/ChiaraPodio/virtual_classroom_system/service/CourseService.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CourseService implements ICourseService {
@@ -23,66 +24,74 @@ public class CourseService implements ICourseService {
     private IStudentService studentService;
 
     @Override
-    public List<Course> getAllCourses() {
+    public List<Course> findAll() {
         return courseRepository.findAll();
     }
 
     @Override
-    public Course getCourseById(Long id) {
-        return courseRepository.findById(id).orElse(null);
+    public Optional<Course> findById(Long id) {
+        return courseRepository.findById(id);
     }
 
     @Override
-    public void saveCourse(Course course) {
+    public void save(Course course) {
         courseRepository.save(course);
     }
 
     @Override
-    public void deleteCourse(Long id) {
+    public void deleteById(Long id) {
+        if(!courseRepository.existsById(id)){
+            throw new RuntimeException("Course not found");
+        }
         courseRepository.deleteById(id);
     }
 
     @Override
-    public void updateCourse(Long course_id, CourseRequestDto courseRequestDto) {
-        Course newCourse = this.getCourseById(course_id);
+    public void update(Long course_id, CourseRequestDto courseRequestDto) {
+        Course newCourse = this.findById(course_id)
+                .orElseThrow(() -> new RuntimeException("Course not found"));
 
         if (courseRequestDto.getName()!= null){
             newCourse.setName(courseRequestDto.getName());
         }
         if (courseRequestDto.getProfessor_id()!=null){
-            newCourse.setProfessor(professorService.getProfessorById(courseRequestDto.getProfessor_id()));
+            newCourse.setProfessor(professorService.findById(courseRequestDto.getProfessor_id()).
+                    orElseThrow(() -> new RuntimeException("Professor not found")));
         }
         if(courseRequestDto.getStudents_id()!=null) {
             List<Student> studentsList = new ArrayList<>();
 
             for (Long studentId : courseRequestDto.getStudents_id()) {
-                Student student = studentService.getStudentById(studentId);
+                Student student = studentService.findById(studentId)
+                        .orElseThrow(() -> new RuntimeException("Student not found"));
                 studentsList.add(student);
             }
 
             newCourse.setStudents(studentsList);
         }
 
-        this.saveCourse(newCourse);
+        courseRepository.save(newCourse);
     }
 
     @Override
-    public void createCourse (CourseRequestDto courseRequestDto) {
+    public Course createCourse (CourseRequestDto courseRequestDto) {
         Course course = new Course();
 
         course.setName(courseRequestDto.getName());
-        course.setProfessor(professorService.getProfessorById(courseRequestDto.getProfessor_id()));
+        course.setProfessor(professorService.findById(courseRequestDto.getProfessor_id()).
+                orElseThrow(() -> new RuntimeException("Professor not found")));
 
         List<Student> studentsList = new ArrayList<>();
 
         for (Long studentId : courseRequestDto.getStudents_id()) {
-            Student student = studentService.getStudentById(studentId);
+            Student student = studentService.findById(studentId)
+                    .orElseThrow(() -> new RuntimeException("Student not found"));
             studentsList.add(student);
         }
 
         course.setStudents(studentsList);
 
-        this.saveCourse(course);
+        return courseRepository.save(course);
     }
 
 }
